requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
awsebcli==3.20.10
blessed==1.20.0
botocore==1.31.85
cachetools==5.3.3
cement==2.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.3
contourpy==1.2.1
cycler==0.12.1
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.5.4
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
image==1.5.33
jmespath==1.0.1
joblib==1.4.0
keras==3.2.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.7
numpy==1.26.3
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pathspec==0.10.1
pillow==10.3.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-environ==0.4.54
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
semantic-version==2.8.5
six==1.16.0
sqlparse==0.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==1.1.0
threadpoolctl==3.4.0
typing_extensions==4.5.0
tzdata==2024.1
urllib3==1.26.18
wcwidth==0.1.9
Werkzeug==3.0.2
wrapt==1.16.0
